12 Policy Emang Mantap

+++ app/Http/Controllers/Screencast/PlaylistController.php
@@ -45,6 +45,8 @@ class PlaylistController extends Controller
 
     public function edit(Playlist $playlist)
     {
+        $this->authorize('update', $playlist);
+
         return view('playlists.edit', [
             'playlist' => $playlist,
             'tags' => Tag::get(),
@@ -53,6 +55,8 @@ class PlaylistController extends Controller
 
     public function update(PlaylistRequest $request, Playlist $playlist)
     {
+        $this->authorize('update', $playlist);
+
         if ($request->thumbnail) {
             Storage::delete($playlist->thumbnail);
             $thumbnail = $request->thumbnail ? $request->file('thumbnail')->store('images/playlist') : $playlist->thumbnail;
@@ -75,6 +79,8 @@ class PlaylistController extends Controller
 
     public function destroy(Playlist $playlist)
     {
+        $this->authorize('delete', $playlist);
+
         Storage::delete($playlist->thumbnail);
         $playlist->tags()->detach();
         $playlist->delete();
